refactor(kitchen pos) : PurgomalumClient 패키지 위치 변경
