+++ src/main/java/kitchenpos/menu/application/external/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain.service;
+package kitchenpos.menu.application.external;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
