+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -5,7 +5,7 @@ import kitchenpos.menu.domain.model.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.product.domain.model.Product;
 import kitchenpos.product.domain.repository.ProductRepository;
-import kitchenpos.menu.domain.service.PurgomalumClient;
+import kitchenpos.menu.application.external.PurgomalumClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
